server/requirement.txt
anyio==3.5.0
asgiref==3.4.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.0.3
cryptography==42.0.8
fastapi==0.72.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.12.0
httplib2==0.22.0
idna==3.3
pdfminer==20191125
pdfminer.six==20231228
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.9.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
sniffio==1.2.0
starlette==0.17.1
tqdm==4.66.4
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.16.0




